reorganize level editor
